rename a stupidly named module
